modules/payt-tenant/requirements.txt
aiopg
unidecode
openpyxl
chardet
py-dateutil
aiomcache
aiohttp
xlrd
xlsxwriter
ckanapi